no message
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@101 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserverweb/source/clientsThread.cpp
@@ -217,7 +217,7 @@ void ClientsTHREAD::clearAllConnections()
 	requestAccess(&connectionWriteAccess,this->id);
 	LPCONNECTION c=connections;
 	LPCONNECTION next=0;
-	for(int i=0;i<nConnections;i++)
+	for(u_long i=0;i<nConnections;i++)
 	{
 		next=c->Next;
 		deleteConnection(c);
